requirements.txt
google-genai
langgraph
numpy
pandas
streamlit
langchain   
typing-extensions
google-generativeai
load_dotenv